Applicatio>DTOs is finished

+++ BlogPost.Domain/Entities/Base.cs
@@ -2,6 +2,6 @@
 
 public class Base
 {
-    public long Id { get; set; } // General UID
+    public int Id { get; set; } // General UID
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow; // General DateTime
 }
